# src/common/client/openapi.json
{
  "openapi": "3.0.2",
  "info": { "title": "Druffle Development API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/admins/pwd": {
      "post": {
        "tags": ["Admins"],
        "summary": "Set Pwd",
        "description": "Set password for administrator through guest token.",
        "operationId": "set_pwd_api_v1_admins_pwd_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NewPasswordIn" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/JwtToken" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
    "/api/v1/hotels/{hotel_id}/messengers": {
      "get": {
        "tags": ["Hotel: Messengers"],
        "summary": "Get Messengers By Hotel Id",
        "description": "Get messengers by hotel id. Applicable for Administrative accounts",
        "operationId": "get_messengers_by_hotel_id_api_v1_hotels__hotel_id__messengers_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelMessengersResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Hotel: Messengers"],
        "summary": "Update Messengers By Hotel Id",
        "description": "Get messengers by hotel id. Applicable for Administrative accounts",
        "operationId": "update_messengers_by_hotel_id_api_v1_hotels__hotel_id__messengers_put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelMessengersUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelMessengersResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "post": {
        "tags": ["Hotel: Messengers"],
        "summary": "Create Messengers",
        "description": "Create new hotel messengers. Applicable for Administrative account",
        "operationId": "create_messengers_api_v1_hotels__hotel_id__messengers_post",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelMessengersUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelMessengersResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
    "/api/v1/hotels/{hotel_id}/contacts": {
      "get": {
        "tags": ["Hotel: Contacts"],
        "summary": "Get Contacts By Hotel Id",
        "description": "Get contacts by hotel id. Applicable for Administrative accounts",
        "operationId": "get_contacts_by_hotel_id_api_v1_hotels__hotel_id__contacts_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Contacts By Hotel Id Api V1 Hotels  Hotel Id  Contacts Get",
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/HotelContactsResponse" }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Hotel: Contacts"],
        "summary": "Update Contacts By Hotel Id",
        "description": "Get contacts by hotel id. Applicable for Administrative accounts",
        "operationId": "update_contacts_by_hotel_id_api_v1_hotels__hotel_id__contacts_put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelContactsUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelContactsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "post": {
        "tags": ["Hotel: Contacts"],
        "summary": "Create Contacts",
        "description": "Create new hotel contacts. Applicable for Administrative account",
        "operationId": "create_contacts_api_v1_hotels__hotel_id__contacts_post",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelContactsUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelContactsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
    "/api/v1/hotels/{hotel_id}/photos/{photo_id}": {
      "put": {
        "tags": ["Hotel: Photos"],
        "summary": "Update Photos By Id",
        "description": "Get photos by hotel id. Applicable for Administrative accounts",
        "operationId": "update_photos_by_id_api_v1_hotels__hotel_id__photos__photo_id__put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": { "title": "Photo Id", "type": "string" },
            "name": "photo_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelPhotosUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelPhotosResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "delete": {
        "tags": ["Hotel: Photos"],
        "summary": "Delete Photos By Id",
        "description": "Delete photos by id. Applicable for Administrative accounts",
        "operationId": "delete_photos_by_id_api_v1_hotels__hotel_id__photos__photo_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": { "title": "Photo Id", "type": "string" },
            "name": "photo_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StatusResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactTypes": {
        "title": "ContactTypes",
        "enum": ["Administrator", "Reception", "Other"],
        "description": "An enumeration."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "HotelContactsResponse": {
        "title": "HotelContactsResponse",
        "required": ["contact_type", "first_name", "last_name", "email", "phone", "hotel_id", "id"],
        "type": "object",
        "properties": {
          "contact_type": { "$ref": "#/components/schemas/ContactTypes" },
          "first_name": { "title": "First Name", "type": "string" },
          "last_name": { "title": "Last Name", "type": "string" },
          "email": { "title": "Email", "type": "string" },
          "phone": { "title": "Phone", "type": "string" },
          "custom_name": { "title": "Custom Name", "type": "string" },
          "hotel_id": { "title": "Hotel Id", "type": "string", "format": "uuid4" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "HotelContactsUpdateRequest": {
        "title": "HotelContactsUpdateRequest",
        "required": ["contact_type", "first_name", "last_name", "email", "phone"],
        "type": "object",
        "properties": {
          "contact_type": { "$ref": "#/components/schemas/ContactTypes" },
          "first_name": { "title": "First Name", "type": "string" },
          "last_name": { "title": "Last Name", "type": "string" },
          "email": { "title": "Email", "type": "string" },
          "phone": { "title": "Phone", "type": "string" },
          "custom_name": { "title": "Custom Name", "type": "string" }
        }
      },
...
      "HotelMessengersResponse": {
        "title": "HotelMessengersResponse",
        "required": ["hotel_id", "id"],
        "type": "object",
        "properties": {
          "telegram": { "title": "Telegram", "type": "boolean", "default": false },
          "whatsup": { "title": "Whatsup", "type": "boolean", "default": false },
          "hotel_id": { "title": "Hotel Id", "type": "string", "format": "uuid4" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "HotelMessengersUpdateRequest": {
        "title": "HotelMessengersUpdateRequest",
        "type": "object",
        "properties": {
          "telegram": { "title": "Telegram", "type": "boolean", "default": false },
          "whatsup": { "title": "Whatsup", "type": "boolean", "default": false }
        }
      },
      "HotelPhotosResponse": {
        "title": "HotelPhotosResponse",
        "required": ["is_active", "url", "preview", "hotel_id", "id"],
        "type": "object",
        "properties": {
          "is_active": { "title": "Is Active", "type": "boolean" },
          "url": { "title": "Url", "type": "string" },
          "preview": { "title": "Preview", "type": "string" },
          "hotel_id": {
            "title": "Hotel Id",
            "anyOf": [{ "type": "string", "format": "uuid4" }, { "type": "string" }]
          },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "HotelPhotosUpdateRequest": {
        "title": "HotelPhotosUpdateRequest",
        "required": ["is_active"],
        "type": "object",
        "properties": { "is_active": { "title": "Is Active", "type": "boolean" } }
      },
...
      "JwtToken": {
        "title": "JwtToken",
        "required": ["access_token", "refresh_token"],
        "type": "object",
        "properties": {
          "access_token": { "title": "Access Token", "type": "string" },
          "refresh_token": { "title": "Refresh Token", "type": "string" }
        }
      },
      "NewPasswordIn": {
        "title": "NewPasswordIn",
        "required": ["password"],
        "type": "object",
        "properties": { "password": { "title": "Password", "type": "string" } }
      },
...
      "StatusResponse": {
        "title": "StatusResponse",
        "required": ["status"],
        "type": "object",
        "properties": {
          "status": { "title": "Status", "enum": ["ok", "fail"], "type": "string" },
          "details": { "title": "Details", "type": "string" }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "JWTokenAuth": { "type": "apiKey", "in": "header", "name": "Authorization" }
    }
  }
}